refactor(models): extract migrate() into models/migrations.rs
Phase 8. Move the 1973-line migrate() function + its column_exists
and reconcile_column_rename helpers + their 5 tests out of
models/mod.rs into models/migrations.rs.

Zero behaviour change. The "idempotent SQL in code, no on-disk
migration files" convention (CLAUDE.md) is preserved — this is pure
code organization, not a migration-strategy change.

Import plumbing:
  - migrations.rs imports the 4 self-migrating sibling modules
    (custom_formats, group_source_map, media_probe_cache,
    nyaa_description_cache) via super::{…} since migrate() calls
    their initialize/seed functions.
  - models/mod.rs declares `pub mod migrations;` and re-exports
    `pub use migrations::migrate;` so every external call site
    (main.rs calls `models::migrate(&db)`) resolves unchanged.

mod.rs:        1994 → 21 lines (pure `pub mod X;` declarations + the
                                 migrate re-export)
migrations.rs: new, 1989 lines

cargo fmt / clippy / build / test all green (601 tests).
